Fix data type conversion for Producto ID

+++ pages/1_🔮_Prediccion_Ordenes_old.py
@@ -60,7 +60,8 @@ def load_ml_objects():
     try:
         meta_product = pd.read_csv(fetch_s3_file_as_bytes("data/fe_products_cost.csv"))
         meta_product = meta_product.rename( columns = {"Referencia interna" : "Producto ID"} )
-        meta_product["Producto ID"] = meta_product["Producto ID"].astype(int).astype(str)
+        meta_product["Producto ID"] = meta_product["Producto ID"].astype(int)
+        meta_product["Producto ID"] = meta_product["Producto ID"].astype(str)
     except Exception as e:
         st.warning("⚠️ No se encontró 'data/fe_products_cost.csv' en AWS S3. Se mostrarán los valores de confianza como 'Desconocido'.")
         st.stop()
